- SurfTile resolution bias bug fixes - LOD algorithm tests and suggestions - Render HUD callback function registration
git-svn-id: svn://mirror.orbiter-radio.co.uk/D3D9client/trunk@547 c0f5cd98-1510-604a-8459-bf7538b8c9c2

+++ Orbitersdk/D3D9Client/D3D9Config.h
@@ -60,7 +60,7 @@ public:
 	int ShadowMapSize;				///< Shadow Map size
 	int CustomCamMode;				///< Custom Camera Mode
 	int TileMipmaps;				///< Enable surface tile mipmaps
-	int LODBias;					///< 3D Terrain resolution bias
+	double LODBias;					///< 3D Terrain resolution bias
 	int MeshRes;					///< Tile patch mesh resolution
 	int TileDebug;					///< Enable tile debugger
 	int TextureMips;				///< Texture mipmap autogen policy
